fix overflow (yuck)

+++ day6a/src/main.rs
@@ -18,18 +18,18 @@ impl Direction {
     }
 }
 
-fn get_guard_position(m: &[Vec<char>]) -> (usize, usize) {
+fn get_guard_position(m: &[Vec<char>]) -> (i64, i64) {
     for (i, row) in m.iter().enumerate() {
         for (j, ch) in row.iter().enumerate() {
             if *ch == '^' {
-                return (j, i);
+                return (j as i64, i as i64);
             }
         }
     }
     panic!("failed to find the guard !!!");
 }
 
-fn move_to(direction: &Direction, from: (usize, usize)) -> (usize, usize) {
+fn move_to(direction: &Direction, from: (i64, i64)) -> (i64, i64) {
     use Direction::*;
     match direction {
         North => (from.0, from.1 - 1),
@@ -39,7 +39,7 @@ fn move_to(direction: &Direction, from: (usize, usize)) -> (usize, usize) {
     }
 }
 
-fn within_bounds(pos: (usize, usize), width: usize, height: usize) -> bool {
+fn within_bounds(pos: (i64, i64), width: i64, height: i64) -> bool {
     (0..width).contains(&pos.0) && (0..height).contains(&pos.1)
 }
 
@@ -47,8 +47,8 @@ fn main() {
     let input = fs::read_to_string("inputs/6.txt").unwrap();
 
     let matrix: Vec<Vec<char>> = input.lines().map(|n| n.chars().collect()).collect();
-    let height = matrix.len();
-    let width = matrix[0].len();
+    let height = matrix.len() as i64;
+    let width = matrix[0].len() as i64;
 
     let mut seen = HashSet::new();
 
@@ -59,7 +59,9 @@ fn main() {
         seen.insert(current_position);
 
         let forward = move_to(&current_direction, current_position);
-        if within_bounds(forward, width, height) && matrix[forward.1][forward.0] == '#' {
+        if within_bounds(forward, width, height)
+            && matrix[forward.1 as usize][forward.0 as usize] == '#'
+        {
             current_direction.rotate();
         } else {
             current_position = forward;
